requirements.txt
pybit
tqdm
mplfinance
pandas
python-telegram-bot
python-dotenv